Sample 2 favorite categories for explore page
