+++ app/components/explore/category-sidebar/general-categories.js
@@ -7,25 +7,12 @@ import { task } from 'ember-concurrency';
 export default Component.extend({
   tagName: '',
   queryCache: service(),
-  userCategories: reads('getUserFavoritesTask.last.value'),
   categories: reads('getCategoriesTask.last.value'),
 
   init() {
     this._super(...arguments);
-    if (get(this, 'session.hasUser')) {
-      get(this, 'getUserFavoritesTask').perform();
-    }
     get(this, 'getCategoriesTask').perform();
   },
-
-  getUserFavoritesTask: task(function* () {
-    return yield get(this, 'queryCache').query('category-favorite', {
-      include: 'category',
-      filter: { user_id: get(this, 'session.account.id') },
-      fields: { category: 'title,slug' },
-      page: { limit: 20 }
-    });
-  }).drop(),
 
   getCategoriesTask: task(function* () {
     return yield get(this, 'queryCache').query('category', {
