+++ app/templates/explore/more.hbs
@@ -1,1 +1,5 @@
-{{explore/explore-more mediaType=mediaType type=type more="explore.more"}}
+{{explore/explore-more
+  getUserFavoritesTask=model.userFavoritesTask
+  mediaType=mediaType
+  type=type
+  more="explore.more"}}
